Add mesh.hpp to house code of unstructured mesh
