+++ CMakeLists.txt
@@ -73,6 +73,7 @@ set(MODMESH_HEADERS
     include/modmesh/base.hpp
     include/modmesh/profile.hpp
     include/modmesh/grid.hpp
+    include/modmesh/mesh.hpp
     include/modmesh/small_vector.hpp
     include/modmesh/ConcreteBuffer.hpp
     include/modmesh/SimpleArray.hpp
